+++ src/components/TextArea/TextArea.stories.tsx
@@ -10,6 +10,7 @@ const meta: Meta<typeof TextArea> = {
   tags: ["autodocs"],
   args: {
     placeholder: "Placeholder",
+    maxLength: 100,
     onChange: action("onChange"),
   },
 }